Рефакторинг, разделение комп. и dto-шек, компоненты в отдельных файлах
