+++ frontend/src/components/registration/dto/second-registration-step.dto.ts
@@ -1,0 +1,13 @@
+import { CreateUserDto } from '#server/common/dto/create-user.dto';
+import { IsOptional } from 'class-validator';
+
+export class SecondStepData extends CreateUserDto {
+  @IsOptional()
+  username;
+
+  @IsOptional()
+  password;
+
+  @IsOptional()
+  email;
+}
